Updated Community page design as per mockup
