+++ components/common/newsletter-subscription/NewsletterSubscription.js
@@ -17,17 +17,45 @@
  */
 
 import * as React from 'react';
-
+import { prefix } from '../../../utils/prefix';
 import styles from './NewsletterSubscription.module.css';
+import Image from 'next-image-export-optimizer';
 
 export default function NewsletterSubscription() {
+
+    const [hoverBtn, setHoverBtn] = React.useState(false);
+
+    let linkArrowPath = prefix + '/images/toc-bg.svg';
+    let linkArrowHoverPath = prefix + '/images/toc-bg-hover.svg';
+
+    const linkArrow = {
+        background: 'url(' + linkArrowPath + ') no-repeat scroll right center',
+        paddingRight: '25px'
+    }
+
+    const linkArrowHover = {
+        background: 'url(' + linkArrowHoverPath + ') no-repeat scroll right center',
+        paddingRight: '25px'
+    }
 
     return (
         <>
             <p>
                 Want to get hand-picked content and the latest news on Ballerina delivered directly to your inbox? Then subscribe to our newsletter today!
             </p>
-            <iframe src="https://resources.wso2.com/l/142131/2022-01-05/b3x14k" width="100%" type="text/html" frameBorder="0" allowTransparency="true" className={styles.iframe}></iframe>
+            <p className={styles.linkWrap}
+                onMouseEnter={() => {
+                    setHoverBtn(true);
+                }}
+                onMouseLeave={() => {
+                    setHoverBtn(false);
+                }}>
+                <a href={`${prefix}/community/ballerina-newsletter`} target="_blank" rel="noreferrer" className={styles.viewAll}>
+                    <span>View all newsletters</span>
+                    {/* <Image src={`${linkArrowPath}`} width={20} height={20} alt="Left Arrow" /> */}
+                </a>
+
+            </p>
         </>
     );
 }
