+++ components/community/get-involved/GetInvolved.js
@@ -89,6 +89,7 @@ export default function GetInvolved(props) {
           </Col>
 
         </Row>
+
       </Container>
     </Col>
   );
