+++ components/community/get-involved/GetInvolved.module.css
@@ -1,13 +1,13 @@
 .getStartLinks {
-    color: #20b6b0 !important;
-    text-decoration: none !important;
-    font-weight: 400;
+  color: #20b6b0 !important;
+  text-decoration: none !important;
+  font-weight: 400;
 }
 
 .getStartLinks:hover {
-    text-decoration: none !important;
-    color: #464646 !important;
-    font-weight: 400;
+  text-decoration: none !important;
+  color: #464646 !important;
+  font-weight: 400;
 }
 
 /* .card {
@@ -17,14 +17,14 @@
 } */
 
 .card h4 {
-    height: 45px;
-    font-size: 22px;
-    color: #57595d !important;
-    margin-top: 20px
+  height: 45px;
+  font-size: 22px;
+  color: #57595d !important;
+  margin-top: 20px;
 }
 
 .customModal {
-    border-radius: 0;
+  border-radius: 0;
 }
 
 /* .customModalBody {
@@ -41,7 +41,7 @@
 } */
 
 .formEmbedded {
-    width: 100%;
-    height: 350px;
-    /* overflow: auto; */
+  width: 100%;
+  height: 350px;
+  /* overflow: auto; */
 }
